+++ tests/TestMaterialShellResponsive.h
@@ -11,6 +11,7 @@ private slots:
     void preservesDestinationAccessAcrossBreakpoints();
     void emitsOnlyOnBreakpointTransitions();
     void appliesVaultPaneContract();
+    void fallbackSearchesAreIndependentAndRestoreFocus();
 };
 
 #endif // KEEPASSXC_TESTMATERIALSHELLRESPONSIVE_H